Move featureFlags so mocks are reloaded upon edit
